+++ src/main/java/school/faang/user_service/controller/recommendation/RecommendationRequestController.java
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import school.faang.user_service.dto.recommendation.RecommendationRequestDto;
-import school.faang.user_service.dto.recommendation.RejectionDto;
 import school.faang.user_service.dto.recommendation.RecommendationRequestFilterDto;
+import school.faang.user_service.dto.recommendation.RejectionDto;
 import school.faang.user_service.exceptions.DataValidationException;
 import school.faang.user_service.service.recommendation.RecommendationRequestService;
