+++ src/main/java/school/faang/user_service/dto/recommendation/RecommendationRequestDto.java
@@ -1,13 +1,9 @@
 package school.faang.user_service.dto.recommendation;
 
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 import school.faang.user_service.entity.RequestStatus;
 
 import java.time.LocalDateTime;
