+++ src/test/java/school/faang/user_service/service/RecommendationRequestServiceTest.java
@@ -1,0 +1,5 @@
+package school.faang.user_service.service;
+
+public class RecommendationRequestServiceTest {
+
+}
