+++ src/test/java/school/faang/user_service/validator/RecommendationRequestValidatorTest.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.validator;
+
+public class RecommendationRequestValidatorTest {
+}
